[#IOPID-2345] Update configuration for Profiles and ServicePreferences
