+++ AnalyticsProfilesChangeFeedInboundProcessorAdapter/index.ts
@@ -1,10 +1,8 @@
 import { QueueClient } from "@azure/storage-queue";
 import { Context } from "@azure/functions";
 
-import * as t from "io-ts";
-
 import { RetrievedProfile } from "@pagopa/io-functions-commons/dist/src/models/profile";
-import { FiscalCode, NonEmptyString } from "@pagopa/ts-commons/lib/strings";
+import { FiscalCode } from "@pagopa/ts-commons/lib/strings";
 
 import { Second } from "@pagopa/ts-commons/lib/units";
 import * as KA from "../outbound/adapter/kafka-outbound-publisher";
@@ -25,21 +23,14 @@ import { profilesAvroFormatter } from "../utils/formatter/profilesAvroFormatter"
 import { httpOrHttpsApiFetch } from "../utils/fetch";
 import { pdvTokenizerClient } from "../utils/pdvTokenizerClient";
 import { createRedisClientSingleton } from "../utils/redis";
-
-export type RetrievedProfileWithMaybePdvId = t.TypeOf<
-  typeof RetrievedProfileWithMaybePdvId
->;
-const RetrievedProfileWithMaybePdvId = t.intersection([
-  RetrievedProfile,
-  t.partial({ userPDVId: NonEmptyString })
-]);
+import { RetrievedProfileWithMaybePdvId } from "../utils/types/decoratedTypes";
 
 const config = getConfigOrThrow();
 
 const profilesConfig = withTopic(
   config.profilesKafkaTopicConfig.PROFILES_TOPIC_NAME,
   config.profilesKafkaTopicConfig.PROFILES_TOPIC_CONNECTION_STRING
-)(config.targetKafka);
+)(config.targetKafkaAuth);
 
 const profilesTopic = {
   ...profilesConfig,
