+++ outbound/adapter/pdv-id-outbound-enricher.ts
@@ -10,10 +10,12 @@ import { Second } from "@pagopa/ts-commons/lib/units";
 import { OutboundEnricher } from "../port/outbound-enricher";
 import { failure, success } from "../port/outbound-publisher";
 import { getPdvId } from "../../utils/pdv";
-import { RetrievedProfileWithMaybePdvId } from "../../AnalyticsProfilesChangeFeedInboundProcessorAdapter";
-import { RetrievedServicePreferenceWithMaybePdvId } from "../../utils/types/decoratedTypes";
+import {
+  RetrievedProfileWithMaybePdvId,
+  RetrievedServicePreferenceWithMaybePdvId,
+  RetrievedUserDataProcessingWithMaybePdvId
+} from "../../utils/types/decoratedTypes";
 import { PdvTokenizerClient } from "../../utils/pdvTokenizerClient";
-import { RetrievedUserDataProcessingWithMaybePdvId } from "../../AnalyticsUserDataProcessingChangeFeedInboundProcessorAdapter";
 
 export type MaybePdvDocumentsTypes =
   | RetrievedServicePreferenceWithMaybePdvId
