+++ utils/__tests__/profilesAvroFormatter.test.ts
@@ -8,10 +8,11 @@ import { ServicesPreferencesModeEnum } from "@pagopa/io-functions-commons/dist/g
 import { NonNegativeInteger } from "@pagopa/ts-commons/lib/numbers";
 import * as avro from "avsc";
 import { profile } from "../../generated/avro/dto/profile";
-import { RetrievedProfileWithMaybePdvId } from "../../AnalyticsProfilesChangeFeedInboundProcessorAdapter";
 import { PushNotificationsContentTypeEnum } from "../../generated/avro/dto/PushNotificationsContentTypeEnumEnum";
 import { ReminderStatusEnum } from "../../generated/avro/dto/ReminderStatusEnumEnum";
 import { aCosmosMetadata } from "../../businesslogic/__mocks__/data.mock";
+import { RetrievedProfileWithMaybePdvId } from "../types/decoratedTypes";
+
 const aUserPDVId = "7f97bf94-6f41-4e8c-97a3-d496125374c0" as NonEmptyString;
 const aRetrievedProfile: RetrievedProfileWithMaybePdvId = {
   ...aCosmosMetadata,
