+++ utils/formatter/deletesAvroFormatter.ts
@@ -2,7 +2,7 @@
 import * as avro from "avsc";
 import { MessageFormatter } from "../kafka/KafkaTypes";
 import { profileDeletion } from "../../generated/avro/dto/profileDeletion";
-import { RetrievedUserDataProcessingWithMaybePdvId } from "../../AnalyticsUserDataProcessingChangeFeedInboundProcessorAdapter";
+import { RetrievedUserDataProcessingWithMaybePdvId } from "../types/decoratedTypes";
 
 // remove me
 export const buildAvroDeleteObject = (
